requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-login
flask-mail
openai
pandas
python-dotenv
colorama
pdfplumber
PyPDF2
reportlab
Pillow
PyMuPDF
stripe
requests
playwright
huey
